Can successfully tar a directory. Writing tests, and working on untarring.
About to do some delete directory things, so making sure I have objects
committed.

Next step: tar/untar from one location to another (with filters), and
tested. Then we'll need to send a tarball to the registry, as well as
retrieve and untar locally.

Aiming for the end of the day.