Phase 1 of proto install created.
Can successfully query the registry for a batch of modules (given either
names, or from a win.json file).

All it does right now is write to a temporary object in the cache
(tar.gz). I'll need to update the registry to also save a properties file
with every version (for quick access without untarring an object).

But, huzzah so far, successful get of a module from a repository (a little
sloppy, but it works).